requirements.txt
Django==3.1.2
django-cors-headers==3.5.0
django-phonenumber-field==5.0.0
django-sendgrid-v5==0.9.0
djangorestframework==3.12.1
phonenumbers==8.12.10